+++ config/scale.js
@@ -17,10 +17,10 @@ module.exports = {
   cluster:    'C1',              // Cluster name to identify it in loadbalancing infrastructure
   cookie:     'node',            // Cookie name for loadbalancing (cookie value will be 'C1'+'N1')
   strategy:   'multiple',
-    // 'single'         - one process (no master and workers)
-    // 'specialization' - multiple processes, one master and different workers for each server (master should not listen ports)
-    // 'multiple'       - multiple processes, one master and identical workers with no sticky (master should listen ports)
-    // 'sticky'         - multiple processes, one master and workers with sticky by IP (master should listen ports)
+    // 'single'   - one process (no master and workers)
+    // 'bundle'   - multiple processes, one master and different workers for each server (master should not listen ports)
+    // 'multiple' - multiple processes, one master and identical workers with no sticky (master should listen ports)
+    // 'sticky'   - multiple processes, one master and workers with sticky by IP (master should listen ports)
 
   workers: api.os.cpus().length, // worker count, e.g. api.os.cpus().length-1 or just number